requirements.txt
pandas==1.5.0
PyYAML==6.0
requests==2.28.1
rich==12.6.0